+++ frontend/pages/admin/makers/index.tsx
@@ -2,7 +2,7 @@ import { Heading, Box, Button } from '@chakra-ui/react';
 import { AddIcon } from '@chakra-ui/icons';
 import { useRouter } from 'next/router';
 import { InternalLayout } from '@/components/layouts/admin/InternalLayout';
-import { MakerStacks } from '@/components/organisms/admin/makers/index/MakerStacks';
+import { MakersStack } from '@/components/organisms/admin/makers/index/MakersStack';
 import { goAdminMakersNew } from '@/utils/routes';
 
 const Index: React.VFC<NoProps> = () => {
@@ -22,7 +22,7 @@ const Index: React.VFC<NoProps> = () => {
           新規登録
         </Button>
       </Box>
-      <MakerStacks />
+      <MakersStack />
     </InternalLayout>
   );
 };
